// state/PrivacyPolicy/privacyPolicy.action-types.ts
export enum ActionTypes {
    POLICY_UPLOAD_REQUEST = 'POLICY_UPLOAD_REQUEST',
    POLICY_UPLOAD_SUCCESS = 'POLICY_UPLOAD_SUCCESS',
    POLICY_UPLOAD_FAIL = 'POLICY_UPLOAD_FAIL',
  
    POLICY_FETCH_REQUEST = 'POLICY_FETCH_REQUEST',
    POLICY_FETCH_SUCCESS = 'POLICY_FETCH_SUCCESS',
    POLICY_FETCH_FAIL = 'POLICY_FETCH_FAIL',
  
    POLICY_DELETE_REQUEST = 'POLICY_DELETE_REQUEST',
    POLICY_DELETE_SUCCESS = 'POLICY_DELETE_SUCCESS',
    POLICY_DELETE_FAIL = 'POLICY_DELETE_FAIL',
  
    POLICY_EDIT_REQUEST = 'POLICY_EDIT_REQUEST',
    POLICY_EDIT_SUCCESS = 'POLICY_EDIT_SUCCESS',
    POLICY_EDIT_FAIL = 'POLICY_EDIT_FAIL',
  
    POLICY_RESET = 'POLICY_RESET',
  }
  